Rework the Repositories UI (#76)
Move the add/edit repository form onto its own dedicated pages, reached by
buttons, so navigating no longer resets a form you're mid-edit on. New
RepoForm component shared by /repos/new and /repos/[id]/edit; the Add
repository button and each row's Edit icon link there.

Widen the list (max-w-4xl -> max-w-6xl) and pack more into each row: an
enabled/disabled indicator icon furthest left with an 'Enabled'/'Disabled'
tooltip, a quick issue-sync toggle switch beside it, and a details line
(default branch, review policy, branch template, label filter, setup-script
presence). The Edit and Delete icon buttons get explicit tooltips.
